"Add basic send and receive functionality using RabbitMQ in separate go files; update docker-compose.yml and go.mod"
This commit adds two new Go source files: `send/send.go` and `recieve/recieve.go`. The `send.go` file contains code for connecting to a RabbitMQ server, declaring a queue, and publishing a message to it. The `recieve.go` file is a stub that currently does not contain any code.

Additionally, the `docker-compose.yml` file has been updated to include a RabbitMQ service based on the `rabbitmq:3.11-management-alpine` image. The environment variables `RABBITMQ_DEFAULT_USER` and `RABBITMQ_DEFAULT_PASS` are set to `guest`. A healthcheck is also included to test if RabbitMQ is running by pinging it. The ports `5672` and `15672` are exposed for AMQP and management access, respectively.

Finally, the `go.mod` file has been updated to include the `github.com/rabbitmq/amqp091-go` package, which is required by the `send.go` file. The Go version is also updated to `go 1.19`.